Praticando o principio de interface em SOLID criando especializações e Factories para atender o desacoplamento da Classe AplicarProva

+++ src/main/java/br/com/ucsal/olimpiadas/Operacao/CadastrarQuestao.java
@@ -14,7 +14,7 @@ public class CadastrarQuestao extends Acao<Questao>{
 			return;
 		}
 
-		Long provaId = f.getEscolherProva().escolherProva(f);
+		Integer provaId = f.getEscolherProva().escolherProva(f);
 		if (provaId == null)
 			return;
